Scrapeモデルをインスタンスメソッドで構成するように戻した
インスタンスメソッドにすることでmachanizeのインスタンス数を減らす
