+++ app/models/article.rb
@@ -5,31 +5,32 @@ class Article < ApplicationRecord
 
   class << self
     def insert_from_tweets(tweets, user_id)
+      crawler = Scrape.new
       article_params = []
       Parallel.each(tweets, in_threads: 10) do |tweet|
         next if tweet.uris.blank?
         article_url = tweet.uris.first.expanded_url.to_s
 
         begin
-          page = Scraper.access_page(article_url)
+          page = crawler.access_page(article_url)
         rescue => e
           next if e.class == Timeout::Error || e.response_code == "404"
           ErrorUtility.log e
         end
 
-        article_params << self.build_params(article_url, page, user_id, tweet)
+        article_params << self.build_params(article_url, crawler, page, user_id, tweet)
       end
 
       Article.insert_all(article_params, unique_by: :url)
     end
 
     private
-      def build_params(article_url, page, user_id, tweet)
+      def build_params(article_url, crawler, page, user_id, tweet)
         now = Time.current
         {
           url:             article_url,
-          title:           Scraper.fetch_title(page),
-          image_meta:      Scraper.fetch_og_image(page),
+          title:           crawler.fetch_title(page),
+          image_meta:      crawler.fetch_og_image(page),
           user_id:,
           tweet_id:        tweet.id,
           tweeted_at:      tweet.created_at,
